Removed PangaVista general code from SocatPangaea
